refactor(decompose): split packages/libraries/graph-tools/bin/vt-graph.ts into folder
