+++ packages/libraries/graph-tools/bin/vt-graph/commands/apply.ts
@@ -1,0 +1,5 @@
+import {graphStateApply} from '../../../src/node'
+
+export async function runApplyCommand(args: readonly string[]): Promise<void> {
+  await graphStateApply(args)
+}
