+++ packages/libraries/graph-tools/bin/vt-graph/commands/mv.ts
@@ -1,0 +1,5 @@
+import {graphMove} from '../../../src/node'
+
+export async function runMvCommand(args: string[]): Promise<void> {
+  await graphMove(0, undefined, args)
+}
