+++ packages/libraries/graph-tools/bin/vt-graph/commands/rename.ts
@@ -1,0 +1,5 @@
+import {graphRename} from '../../../src/node'
+
+export async function runRenameCommand(args: string[]): Promise<void> {
+  await graphRename(0, undefined, args)
+}
