+++ packages/libraries/graph-tools/bin/vt-graph/commands/structure.ts
@@ -1,0 +1,5 @@
+import {runStructureCommand} from '../../structureCommand'
+
+export async function runStructureCliCommand(args: string[]): Promise<void> {
+  await runStructureCommand(args)
+}
